Separated domain objects, typehints madness
